# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SFLS Server
  version: 0.0.0

paths:
  /libraries:
#    get:
#      tags:
#        - libraries
#      summary: Returns a list of library items
#      responses:
#        "200":


    post:
      tags:
        - libraries
      summary: Create a library item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                path:
                  type: string
                mediaType:
                  type: string
                  enum: [movie, tvshow]
              required:
                - name
                - path
                - mediaType
      responses:
        "200":
          description: Single library item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  path:
                    type: string
                  name:
                    type: string
                  mediaType:
                    type: string
                required:
                  - id
                  - path
                  - name
                  - mediaType
        "500":
          description: Internal server error

  /media:
    get:
      tags:
        - media
      summary: Returns a list of media items
      responses:
        "200":
          description: List of media items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
        "500":
          $ref: "#/components/responses/Problem"

  /media/{mediaId}:
    get:
      tags:
        - media
      summary: Returns a single media item
      parameters:
        - in: path
          name: mediaId
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Single media item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"

components:
  responses:
    Problem:
      description: Unexpected problem
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
        - title
        - status
    Media:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        title:
          type: string
        season:
          type: integer
        episode:
          type: integer
        attributes:
          type: object
          additionalProperties: true
      required:
        - id
        - created_at
        - updated_at
        - title
        - attributes